Refactor integration tests for context menu
Merge the integration tests for the configuration and the integration tests for the functionality and split the tests by attribution view, audit view and all views. This slightly decreases the test coverage, but also decreases the runtime of the tests and makes them easier to understand and update.
